requirements.txt
streamlit==1.54.0
google-generativeai==0.8.6
databricks-sql-connector==4.2.5
pandas==2.2.0
openpyxl==3.1.2
boto3==1.28.0
google-cloud-bigquery==3.12.0
google-auth==2.23.0
requests==2.31.0
